requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
httpx>=0.24.0
google-cloud-bigquery>=3.11.0
pydantic-settings>=2.0.0
